third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-image-computed.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-image value 'conic-gradient(red 0deg, gold 1turn)'
  assert_equals: expected "conic-gradient(rgb(255, 0, 0) 0deg, rgb(255, 215, 0) 360deg)" but got "conic-gradient(rgb(255, 0, 0) 0deg, rgb(255, 215, 0) 1turn)"
Harness: the test ran to completion.
